+++ packages/firebase_admob/example/android/app/src/main/java/io/flutter/plugins/firebaseadmobexample/MainActivity.java
@@ -8,20 +8,21 @@ import android.os.Bundle;
 import android.widget.TextView;
 import com.google.android.gms.ads.formats.UnifiedNativeAd;
 import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin;
 import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin;
 
 public class MainActivity extends FlutterActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     GeneratedPluginRegistrant.registerWith(this);
-    FirebaseAdMobPlugin.setNativeAdGenerator((UnifiedNativeAd ad) -> {
-      final TextView myAdView = new TextView(this);
-      myAdView.setWidth(200);
-      myAdView.setHeight(200);
-      myAdView.setText("This is an ad!");
-      return myAdView;
-    });
+    FirebaseAdMobPlugin.setNativeAdFactory(
+        (UnifiedNativeAd ad) -> {
+          final TextView myAdView = new TextView(this);
+          myAdView.setWidth(200);
+          myAdView.setHeight(200);
+          myAdView.setText("This is an ad!");
+          return myAdView;
+        });
   }
 }