+++ packages/firebase_admob/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.9.0+8
+
+* Remove AndroidX warning.
+
 ## 0.9.0+7
 
 * Update Android gradle plugin, gradle, and Admob versions.